Moved timepicker to components
