+++ public/app/features/dashboard/components/TimePicker/index.ts
@@ -1,0 +1,1 @@
+export { TimePickerCtrl } from './TimePickerCtrl';
