+++ public/app/features/dashboard/components/TimePicker/settings.html
